Build of 3d logic moved to RealSpaceCanvas.
